requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
sqlite3  # Built-in with Python

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client and utilities
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4

# Medical imaging
pydicom==2.4.4
nibabel==5.2.0

# Machine Learning
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4

# Explainability
captum==0.6.0
matplotlib==3.8.2
seaborn==0.13.0

# Background tasks and queue
celery[redis]==5.3.4
redis==5.0.1

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# File handling and validation
python-magic==0.4.27
python-magic-bin==0.4.14

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.27.0
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Deployment
gunicorn==21.2.0

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
rich==13.7.0
typer==0.9.0
